vps_package/requirements.txt
coinbase-advanced-py
streamlit
pandas
plotly
urllib3<2.0.0
